Removes the MOTD from the chat window and pops it up when you join the server in a brand new window instead. This window must be closed to continue to the normal menu, and basically consists of the MOTD + Rules + Map info + some useful links. Various config changes etc. made as a result, examples should be provided. Doesn't include actual configs or MOTD I tested with. When/if this goes through I'll throw them Zuh's way or whatever.
A global list (ew I know) is used to check which ckeys have pressed the button so you only see it once per round. It's also available as a Server Information verb in the OOC tab.

The idea behind this is that it can be used to show current info, event info, map-specific info, or whatever, as well as preventing "I didn't read the rules" as an excuse because you literally have to click a button on a page showing you the rules to continue.