Removes some ladder requirements
Removes the requirement for ladders to be on open space to correctly link with the ladders below, and stops them adding lattice to the tile they're on. Ladder pairs should now function on any terrain they're placed on.
This allows ladders to be used between 2 Z-levels that have no open space and no default base turfs of open space, so you can have Z-levels which are ladder-traversable whilst still being further apart than just sat on top of each other.