Makes the info window/new player panel a bit more robust
Removes 1 call to show the new player panel which I'm certain is defunct and should not be occurring (on pref setup), to prevent the panel appearing when the player still has the server info window open.
Forcibly closes the new player panel when the client logs in, to ensure that it cannot be open (it would remain open if it already was and the player was connected over the server restart, at which point it would break).
Adds a global definition of the player panel window ID so that we're sure we're creating and closing the same window ID.

Suggestions of other ways to do this very welcome.